moved some methods from operation_node.py to list.py; small fixes
